fix(jwt-impl): Remove OnlyStringKeys type constraint from ValidCustomClaims
This commit removes the OnlyStringKeys type constraint from the ValidCustomClaims type definition. Previously, ValidCustomClaims was designed to ensure that all keys in the custom claims object are strings, using the OnlyStringKeys type. However, upon further consideration, this constraint was found to be unnecessary for our use case. JavaScript treats numeric keys and their string equivalents interchangeably in objects (e.g., an object's key '0' is treated the same as 0). This realization led to the decision that enforcing keys to be strictly string types added unnecessary complexity without providing significant benefits.